Port to Kyber v3, remove references to onet
