+++ example/go.mod
@@ -1,0 +1,10 @@
+module go.dedis.ch/purbs/example
+
+go 1.13
+
+require (
+	go.dedis.ch/kyber/v3 v3.0.12
+	go.dedis.ch/purbs v0.0.0-00010101000000-000000000000
+)
+
+replace go.dedis.ch/purbs => ../purbs
