+++ experiments-encoding/go.mod
@@ -1,0 +1,13 @@
+module go.dedis.ch/purbs/experiments-encoding
+
+go 1.13
+
+require (
+	github.com/nikirill/go-crypto v0.0.0-20180601122125-13016ed4f5b3
+	github.com/stretchr/testify v1.5.1
+	go.dedis.ch/kyber/v3 v3.0.12
+	go.dedis.ch/purbs v0.0.0-20200113134750-f375b0e26cef
+	gopkg.in/urfave/cli.v1 v1.20.0
+)
+
+replace go.dedis.ch/purbs => ../purbs
