+++ experiments-encoding/pgp/pgp_test.go
@@ -1,12 +1,12 @@
 package pgp
 
 import (
+	"fmt"
+	"log"
 	"testing"
 
-	"fmt"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"gopkg.in/dedis/onet.v1/log"
 )
 
 func TestNewPGP(t *testing.T) {
@@ -27,7 +27,7 @@ func TestPGP_EncryptDecrypt(t *testing.T) {
 	fmt.Println("=================TEST Normal PGP=================")
 	enc, err := Encrypt(msg, recipients, false)
 	if err != nil {
-		log.ErrFatal(err)
+		log.Fatal(err)
 	}
 	//fmt.Printf("Encryption:\n%s\n", sender.ArmorEncryption(enc))
 	dec, err := recipients[len(recipients)-1].Decrypt(enc)
@@ -40,7 +40,7 @@ func TestPGP_EncryptDecrypt(t *testing.T) {
 	fmt.Println("==========TEST PGP with Hidden Recipients========")
 	enc, err = Encrypt(msg, recipients, true)
 	if err != nil {
-		log.ErrFatal(err)
+		log.Fatal(err)
 	}
 	//fmt.Printf("Encryption:\n%s\n", ArmorEncryption(enc))
 	dec, err = recipients[len(recipients)-1].Decrypt(enc)
