+++ purbs/padding_test.go
@@ -1,8 +1,9 @@
 package purbs
 
 import (
-	"github.com/stretchr/testify/require"
 	"testing"
+
+	"github.com/stretchr/testify/require"
 )
 
 func TestPad(t *testing.T) {
